+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
 
+## [0.16.3] - 2025-01-02
+
 ### Added
 - `pw.io.iceberg.write` method for writing Pathway tables into Apache Iceberg.